requirements.txt
beautifulsoup4==4.12.2
blurhash==1.1.4
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
idna==3.4
Mastodon.py==1.8.1
oauthlib==3.2.2
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
tweepy==4.14.0
urllib3==2.0.3
